Refactor to use Event-level API, not Promise
Using the Promise API has a large problem: we need to be able to destroy promises in their own .then() continuations. Using the Event-level API allows us to sidestep that issue by making CoroutineAdapter an Event that can be woken up -- the CoroutineAdapter will outlive all intervening continuations. This restricts us to co_awaiting kj::Promises inside functions that themselves return kj::Promises, but that's a reasonable restriction for a first draft.